Добавлены репозитории и сервисы.

+++ src/main/java/com/example/application/data/service/DemandRepository.java
@@ -5,6 +5,6 @@ import com.example.application.data.entity.Demand;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 
-public interface DemandRepository extends JpaRepository<Demand, Integer> {
+public interface DemandRepository extends JpaRepository<Demand, Long> {
 
 }
